+++ packages/junior/tests/unit/runtime/thread-context.test.ts
@@ -1,5 +1,9 @@
 import { describe, expect, it } from "vitest";
-import { getAssistantThreadContext } from "@/chat/runtime/thread-context";
+import {
+  getAssistantThreadContext,
+  getTeamId,
+} from "@/chat/runtime/thread-context";
+import { runWithWorkspaceTeamId } from "@/chat/slack/workspace-context";
 
 describe("getAssistantThreadContext", () => {
   it("uses the current raw message ts for the first non-DM thread reply", () => {
@@ -59,5 +63,43 @@ describe("getAssistantThreadContext", () => {
         threadId: "slack:D12345:1700000000.300",
       } as any),
     ).toBeUndefined();
+  });
+});
+
+describe("getTeamId", () => {
+  it("uses the raw Slack workspace team when Slack provides it", () => {
+    expect(
+      getTeamId({
+        raw: {
+          team_id: "T_RAW",
+        },
+      } as any),
+    ).toBe("T_RAW");
+  });
+
+  it("falls back to the inbound webhook workspace team", async () => {
+    await runWithWorkspaceTeamId("T_WORKSPACE", async () => {
+      await Promise.resolve();
+      expect(
+        getTeamId({
+          raw: {
+            channel: "C12345",
+            ts: "1700000000.200",
+          },
+        } as any),
+      ).toBe("T_WORKSPACE");
+    });
+  });
+
+  it("prefers the inbound workspace over a Slack Connect author team", () => {
+    runWithWorkspaceTeamId("T_WORKSPACE", () => {
+      expect(
+        getTeamId({
+          raw: {
+            user_team: "T_EXTERNAL",
+          },
+        } as any),
+      ).toBe("T_WORKSPACE");
+    });
   });
 });
